Add a node_user var, replace ansible_user_id
Targeting the ansible_user_id var means node can only be installed for
the operator when the oparator runs the role. This change adds a
'node_user' var which will default to the ansible_user_id if a
'target_user' is not specified. Thus an administrative
operator/ansible_user will be able to install node for other accounts
besides their own, e.g. when said admin needs to push node for a
developer.