Remove package install and become directive
Since the role is intended for the user space and might be used by a
user without sudo, the package install task and become directives were
removed. README.md was updated to mention the packages as requirements,
and to advise admins to use "become" with the role when installing
NodeJS for another user.